+++ geoprox-core/benches/spatial_index_benchmark.rs
@@ -1,7 +1,7 @@
 use criterion::{
     black_box, criterion_group, criterion_main, BatchSize, BenchmarkId, Criterion, Throughput,
 };
-use geoprox_core::cache::SpatialIndex;
+use geoprox_core::{cache::SpatialIndex, models::LatLngCoord};
 use rand::prelude::*;
 
 const UNIT_CAPACITY: i32 = 10;
@@ -12,6 +12,7 @@ const CAPACITY_RANGE: [i32; 5] = [
     UNIT_CAPACITY * 1000,
     UNIT_CAPACITY * 10_000,
 ];
+const MAX_COUNT: usize = 100;
 
 fn random_geohash(rng: &mut ThreadRng) -> String {
     geohash::encode(
@@ -19,6 +20,17 @@ fn random_geohash(rng: &mut ThreadRng) -> String {
         SpatialIndex::DEFAULT_DEPTH,
     )
     .unwrap()
+}
+
+fn random_query(rng: &mut ThreadRng) -> (LatLngCoord, f64) {
+    let ((lat, lng), range) = (
+        (
+            rng.gen_range(-90f64..90f64).into(),
+            rng.gen_range(-180f64..180f64).into(),
+        ),
+        rng.gen_range(50..500).into(),
+    );
+    ([lat, lng], range)
 }
 
 fn seed_index(size: &i32, rng: &mut ThreadRng) -> SpatialIndex<String> {
@@ -31,8 +43,12 @@ fn seed_index(size: &i32, rng: &mut ThreadRng) -> SpatialIndex<String> {
 
 fn insert_benchmark(c: &mut Criterion) {
     let mut group = c.benchmark_group("spatial_index/insert");
-
+    let mut rng = rand::thread_rng();
     for capacity in CAPACITY_RANGE.iter() {
+        let mut data: Vec<(String, String)> = Vec::new();
+        for n in 0..*capacity {
+            data.push((n.to_string(), random_geohash(&mut rng)));
+        }
         group.throughput(Throughput::Elements(*capacity as u64));
         group.bench_with_input(
             BenchmarkId::from_parameter(capacity),
@@ -40,15 +56,10 @@ fn insert_benchmark(c: &mut Criterion) {
             |b, &capacity| {
                 b.iter_batched(
                     || {
-                        let mut rng = rand::thread_rng();
                         let geo_index = SpatialIndex::new(capacity as usize);
-                        let mut data: Vec<(String, String)> = Vec::new();
-                        for n in 0..capacity {
-                            data.push((n.to_string(), random_geohash(&mut rng)));
-                        }
-                        (black_box(geo_index), data)
+                        black_box(geo_index)
                     },
-                    |(mut geo_index, data)| {
+                    |mut geo_index| {
                         data.iter()
                             .for_each(|(key, ghash)| geo_index.place_resource(key, ghash))
                     },
@@ -64,22 +75,21 @@ fn query_range_benchmark(c: &mut Criterion) {
     let mut group = c.benchmark_group("spatial_index/range_query");
 
     for capacity in CAPACITY_RANGE.iter() {
+        let mut rng = rand::thread_rng();
+        let geo_index: SpatialIndex<String> = black_box(seed_index(capacity, &mut rng));
+
         group.throughput(Throughput::Elements(*capacity as u64));
         group.bench_with_input(BenchmarkId::from_parameter(capacity), capacity, |b, _| {
             b.iter_batched(
-                || {
-                    let mut rng = rand::thread_rng();
-                    let geo_index: SpatialIndex<String> = seed_index(capacity, &mut rng);
-                    let ((lat, lng), range) = (
-                        (
-                            rng.gen_range(-90f64..90f64).into(),
-                            rng.gen_range(-180f64..180f64).into(),
-                        ),
-                        rng.gen_range(50..500).into(),
-                    );
-                    (geo_index, (lat, lng), range)
-                },
-                |(geo_index, (lat, lng), range)| geo_index.search([lat, lng], &range, None),
+                || random_query(&mut rng),
+                |([lat, lng], range)| geo_index.search([lat, lng], range, MAX_COUNT, false, None),
+                BatchSize::SmallInput,
+            );
+        });
+        group.bench_with_input(BenchmarkId::new("sorted", capacity), capacity, |b, _| {
+            b.iter_batched(
+                || random_query(&mut rng),
+                |([lat, lng], range)| geo_index.search([lat, lng], range, MAX_COUNT, true, None),
                 BatchSize::SmallInput,
             );
         });